Replaced old density.py file with height.py
